Update stream header proposal based on 1/9 Authors' call
1. Rebased all commits on main, updated text to match
2. Remove proposals 1 and 2
3. Add Track Alias back to all OBJECT formats, we now have both Subscription ID and Track Alias.

This is redundant, the receiver can determine the track name and namespace from the Susbscription ID alone, but the group is unsure if we need Subscription ID in place in the long term.  It was added based on the previous call, when there was concern about mixing objects from different subscriptions on the same stream.The plan is to have both for now, and let implementation experience and continued discussion guide whether we remove one or the other.