Delivery Preference for VOD and reliable live.
The default behavior for live streams will be `prefer-skip` as encoded
in the priority field. This doesn't work for VOD and reliable live, as
they want to receive objects in creation order instead.

This is a counter-proposal to the FETCH keyword which attempts to
accomplish something similar, but in a way that won't work for reliable
live.

This PR also removes relative start/end offsets because SUBSCRIBE was
getting too complex. An application that wants to start N groups back
needs to use a timeline to learn about the absolute group/object IDs.