Remove normative restriction on relays forwarding decision input
I removed this sentence:

Relays MUST NOT depend on OBJECT payload content for making forwarding decisions and MUST only depend on the fields, such as priority order and other metadata properties in the OBJECT message header.

We already have text explaining that the object payload content can be encrypted (section 2.1).  I think that makes it clear that a relay can't do anything with the object payload unless it has the encryption keys or a priori knowledge they are cleartext.

The sentence itself was overly restrictive (relays could use other inputs, for example out of band information from configuration or business relationships).  The preceding sentence explains how to find subscribers to this track, the subsequent one says "MUST forward", and we have an entire section explaining what to do with priority order now.

Fixes: #217