Rewrite the Priority section to improve clarity
This change provides a step-by-step algorithm for how to pick the next
object to send, and which priorities take precedence over another, all
in one place.

This PR may or may not change the normative behavior of priorities
depending on one's interpretation of the current text (which, as we
learned in today's call, appears to vary between people).

Fixes #517