Clarify SUBSCRIBE now that FETCH is landed
Changes SUBSCRIBE so you can't go back farther than the current Group with any of the filter modes.  If a SUBSCRIBE tries to go back farther, it is closed with a SUBSCRIBE_OK.

Alternative, instead of failing the subscription, we could specify that the SUBSCRIBE will be truncated and start at the beginning of the current group.