Adding Absolute Joining Fetch (#747)
This PR adds a Absolute Joining Fetch, in which the Start Group is
defined absolutely instead of relatively as is done with the existing
Relative Joining Fetch.

An Absolute Joining Fetch assists a client in making an optimal ABR
switch, as described by this
[presentation](https://docs.google.com/presentation/d/1YPLTpHjrJpNUtjvw58fAvE2NhiWHjuptDwTcjaRvVcI/edit#slide=id.g331a8043eb5_0_156)
delivered during the Denver Interim.

Fixes #730

---------

Co-authored-by: ianswett <ianswett@users.noreply.github.com>