Add error codes for Duplicate and Invalid Subscribe ID (#822)
Duplicate Subscribe ID is a session error, to mirror Duplicate Track
Alias.

Invalid Subscribe ID is a FETCH_ERROR (not a session error) for Joining
fetch if the dependent ID does not exist. Since the client pipelines
Subscribe+Joining Fetch, we don't want to close the session if the
Subscribe fails and invalidates the ID.

Fixes: #674