Harmonizing serialization of Parameters and Extensions. (#761)
This PR applies extends serialization of Extensions (in which even types
carry a varint payload and odd types carry a Length/Value payload) to
parameters. The serialization is defined first for parameters and then
referenced for Extensions.

The DELIVERY TIMEOUT parameter changes from a type of 0x03 to 0x02 as it
carries a varint payload.

The AUTHORIZATION INFO parameter changes from 0x02to 0x03 as it carries
a TL payload.

Renumbers the CLIENT_SETUP and SERVER_SETUP messages to make supporting
both serializations easier.

No changes in type required for the SETUP parameters. 

Fixes #672

---------

Co-authored-by: ianswett <ianswett@users.noreply.github.com>
Co-authored-by: afrind <afrind@users.noreply.github.com>
Co-authored-by: Suhas Nandakumar <snandaku@cisco.com>