Explain interaction of caching, status and extensions (#826)
This is the simplest proposal, but doesn't cover the case of an
extension that was added by the Original Publisher that was
added/modified/removed by a relay illegally (eg: LoC metadata). To
support that, the extension mechanism needs a way to identify immutable
extensions.

Also added a normative statement about forwarding the same object more
than once.

Fixes: #741 
Fixes: #745

---------

Co-authored-by: ianswett <ianswett@users.noreply.github.com>
Co-authored-by: Martin Duke <martin.h.duke@gmail.com>