Unify parameter parsing by always using byte length
Instead of the number of parameters, as are used in some control messages.

I thought the parsers were already the same, but it turns out some spots use number of parameters and Objects use total length.  I don't think that difference was intentional, but let me know if it was.