Minimal Stream definition (#1020)
I believe this is the minimal change needed to address #1001.

I’m unsure about the preferred ordering of terms. Is there any guidance
on that, or on other editorial details of this kind?

---

Section 9. Talks about “Data streams”. And it is referenced in a couple
of places as “data stream”. Without a prior definition and a
differentiation of “Data stream” and “streams”.
I don’t know the best way forward there. If even a tangible difference
between Data stream and stream exists in the context of this document,
or what the original reason for the “data” stream term was.

Note. [RTP over
QUIC](https://www.ietf.org/archive/id/draft-ietf-avtcore-rtp-over-quic-14.html#section-2-11.18.1)
also “struggles” with the usage and definition of the word stream.

This is my first PR here. I am sorry if it ends up being more of a
burden than a help. I know this borders on bikeshedding, but hopefully
it nudges things a little forward and improves clarity for new readers.