OBJECT_DATAGRAM_STATUS (#1082)
Fixes: #1076 

I think Object Datagram Status might not have been updated with the
other values moved? Added a table and altered the text to reflect the
values given above.

Maybe this is worth another issue, but I see that extensions being
declared as present with a length of 0 is sometimes a protocol violation
and sometimes not. (Datagram vs Stream). Not sure that's intentional?

---------

Co-authored-by: afrind <afrind@users.noreply.github.com>