Security considerations on Relay: ANNOUNCE and SUBSCRIBE_NAMESPACE considerations (#1095)
If an endpoint sends a relay a large number of ANNOUNCE or
SUBSCRIBE_NAMESPACE messages, the relay would be required to store the
mappings, potentially forever. A flood of bogus namespaces could bloat
data structures on a relay that store namespace tables. This PR adds
some language in the "Security considerations" section in order to
encourage implementations to guard against this.

Resolves #1094

---------

Co-authored-by: afrind <afrind@users.noreply.github.com>