Explain Modularity of MOQT (#1107)
Fixes: #908

Added a new term "intermediary" which was our old Relay definition, and
redefined Relay to mean "an intermediary that follows all the rules". An
example of a non-relay intermediary might be an ingest proxy, that only
supports PUBLISH related messages upstream and downstream.

Added a top-level "Modularity" section between "Sessions" and
"Publishing and Receiving Tracks" - open to suggestions if there's a
better place.