Explain FETCH routing for relays (#1117)
Fixes: #728

* Use the same rules as SUBSCRIBE to find available publishers
---
EDIT: 

Originally this PR had
* MUST send to at least one, and try against others for
UNKNOWN_STATUS_IN_RANGE

But now it reads:

* MAY send to any publisher if there is more than one.
---

This leaves the door open for making the FETCHes in parallel, but
doesn't require it.

I didn't see any other FETCH_ERROR codes that seemed retryable in this
way, but if there are we can make the relay language more generic.

REMINDER for Editors: This is targeted against #1116, retarget main
before merging.