RFC: Generic Filters for Subscriptions and FETCH
Based on proposal in #1068

One thing that is strange is allowing Group ID filters here, since these overlap with the built-in filters in Subscribe and Fetch.  This may be what Victor was suggesting in Stockholm.

The pathological case of this encoding is for every other object, which would double the length of the Values array.  We could fix it by shifting the start value left by one and using the LSB to indicate single vs range, but it seemed like overkill.