add marker for end to end header authentication (#1025)
Add a header extension that indicates extensions after this can be end
to end authenticated ( not encrypted ).

This PR is to set up for work in ObjCrypto draft that provided the end
to end encryption. This is becoming an implementation blocker for us as
we start working through the end to end security stuff.

Related to #748

---------

Co-authored-by: Suhas Nandakumar <snandaku@cisco.com>
Co-authored-by: afrind <afrind@users.noreply.github.com>