Fetch Object serialization optimization (#949)
This PR adds a 1-byte Serialization field to the FETCH object definition
and defines 4 code points:

0x0 - status quo - Group/Subgroup and Object ID fields are present. This
would be used at the start of a track and adds 1 byte to the current
serialization.

0x1 - The Group ID and Subgroup ID of this object are identical to those
of the prior object and Object ID increments by 1. The Group ID,
Subgroup ID and Object ID fields are not present. This code point would
most commonly occur in a FETCH response and would save 2 bytes per
object.

0x2 - The Group ID field exists and Subgroup ID and Object ID are both
0. The SubGroup ID and Object ID fields are not present. This code point
would be used at the start of each new Group and saves 1 byte per
object.

0x3 - The Group ID of this object is identical to that of the prior
Object. Subgroup ID increments by 1 and Object ID increments by 1. The
Subgroup ID and Object ID fields are not present. This code point would
be used at the start of a new Subgroup and would save 1 byte per object.


Fixes #660

---------

Co-authored-by: ianswett <ianswett@users.noreply.github.com>
Co-authored-by: afrind <afrind@users.noreply.github.com>
Co-authored-by: Ye-Kui Wang <98077873+yekuiwang@users.noreply.github.com>
Co-authored-by: Suhas Nandakumar <snandaku@cisco.com>