Parameters for Group Order, Subscribe Priority and Subscription Filter (redo) (#1273)
Note: this is a redo of #1210, targeted against main

This is a proposal based on the most recent discussion of parameters and
extensibility.

It moves Group Order and Subscriber Priority to parameters in SUBSCRIBE,
SUBSCRIBE_UPDATE, PUBLISH, PUBLISH_OK. The PR specifies a value for the
receiver to use if the parameter is omitted.

It defines a structure for a Subscription Filter as it exists today.  
* The wire format for SUBSCRIBE and PUBLISH_OK use this directly (not as
a param, no wire format change)
* SUBSCRIBE_UPDATE is changed to accept a parameter containing this
filter when the subscriber wants to change it

Items of note:

1. This adds a new filter type "No Filter", addressing #1035.
2. Relative filter types are only allowed in SUBSCRIBE, because they
require the SUBSCRIBE_OK containing Largest to communicate what actually
happened, but maybe this could be relaxed?
3. The values used by Group Order are different depending on the
message. This preserves existing behavior while eliminating the
"Default" wire value previously allowed in SUBSCRIBE.
4. Forward is retained as a Fixed field, rather than in all messages
where it appears. This is based on the fact that there's no sensible
"common" value.
5. Other than Forward, everything that can be changed in
SUBSCRIBE_UPDATE is now a parameter, and omitted means "unchanged".

Fixes: #872 
Fixes: #878 
Fixes: #1035 
Closes: #1043
Closes: #1044 
Closes: #1059

---------

Co-authored-by: Ye-Kui Wang <98077873+yekuiwang@users.noreply.github.com>
Co-authored-by: ianswett <ianswett@users.noreply.github.com>