Allow unknown ranges in a FETCH response (#1331)
This is a proposal to address several open FETCH issues.

Replace UNKNOWN_STATUS_IN_RANGE with a data-stream encoding of an
UNKNOWN range. This requires leaving DOES_NOT_EXIST serializations in
fetch, so a omitted range can be split between DNE and UNKNOWN. But that
seems OK to me.

I'm unsure if it addresses #945 completely, but it gives a bandwidth
constrained publisher another option that doesn't completely fail the
fetch. (Ian believes it does resolve #945, unless others object)

Fixes: #1265 
Fixes: #1057 
Fixes: #945

---------

Co-authored-by: ianswett <ianswett@users.noreply.github.com>
Co-authored-by: Suhas Nandakumar <snandaku@cisco.com>